Split TPM LUKS container activation out of crypt.go
This moves the TPM-specific APIs out of crypt.go and temporarily
into crypt_tpm.go, before the TPM code moves to a subpackage.
